# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ninja Kilat WMS",
    "description": "Warehouse management for Ninja Van's GrabMart Kilat fulfilment: inbound and putaway, guided picking, weekly stock opname, barcode and license-plate registration, and an isolated training mode.",
    "version": "0.1.0",
    "contact": {
      "name": "Ninja Van — Kilat fulfilment"
    }
  },
  "paths": {
...
    "/api/replenishment": {
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "Open moves from overflow to the pick face, most urgent first",
        "description": "Open moves from overflow to the pick face, most urgent first.\n\nA pick face at zero outranks one merely low: the first stops picking, the\nsecond only threatens to.",
        "operationId": "list_replenishment_api_replenishment_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplenishmentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/replenishment/{task_id}/complete": {
      "post": {
        "tags": [
          "flow"
        ],
        "summary": "Book the physical move: units leave overflow and land on the pick face",
        "description": "Book the physical move: units leave overflow and land on the pick face.\n\nTwo ledger movements, not one. A relocation that wrote a single row would\nlose which location the stock left, and the rack map would be wrong until the\nnext count.",
        "operationId": "complete_replenishment_api_replenishment__task_id__complete_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplenishDoneIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/restock": {
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "List Restock",
        "operationId": "list_restock_api_restock_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestockList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/restock/{request_id}/send": {
      "post": {
        "tags": [
          "flow"
        ],
        "summary": "A supervisor confirms the quantity and sends the request to the hub",
        "description": "A supervisor confirms the quantity and sends the request to the hub.",
        "operationId": "send_restock_api_restock__request_id__send_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          },
          {
            "name": "qty",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Qty"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transfers": {
      "post": {
        "tags": [
          "flow"
        ],
        "summary": "Dispatch a tote from the hub to a darkstore",
        "description": "Dispatch a tote from the hub to a darkstore.\n\nThe dispatched quantity is what makes the receiving end checkable even when\nthe brand sent no manifest — this number is ours, so hop 2 of the chain of\ncustody always has an authority (PRD 5.6). Stock leaves the hub's ledger now;\nit enters the darkstore's when someone receives it.",
        "operationId": "create_transfer_api_transfers_post",
        "parameters": [
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "Transfers touching this site, in either direction",
        "description": "Transfers touching this site, in either direction.",
        "operationId": "list_transfers_api_transfers_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pos/outbox": {
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "What the boundary is doing",
        "description": "What the boundary is doing — or, during the pilot, deliberately not doing.\n\nThe most valuable thing on this endpoint is `push_enabled`. A supervisor\nlooking at a healthy-looking queue must be able to tell that nothing is being\nsent on purpose, rather than concluding the integration works.",
        "operationId": "outbox_health_api_pos_outbox_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboxHealth"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shortfalls": {
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "Declared shortfalls, and who declared them",
        "description": "Declared shortfalls, and who declared them.\n\nA short pick corrects stock without a signature, so the only control on it is\nthat someone reviews these — and that a person declaring them far more often\nthan their colleagues is visible (PRD 8.9.5).",
        "operationId": "list_shortfalls_api_shortfalls_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Site Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-forwarded-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortfallList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "Ok"
      },
...
      "OutboxHealth": {
        "properties": {
          "push_enabled": {
            "type": "boolean",
            "title": "Push Enabled",
            "description": "False means shadow mode: every number computed, nothing sent"
          },
          "lanes": {
            "items": {
              "$ref": "#/components/schemas/OutboxLane"
            },
            "type": "array",
            "title": "Lanes"
          },
          "oldest_pending_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Pending Seconds"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "push_enabled",
          "lanes",
          "note"
        ],
        "title": "OutboxHealth"
      },
...
      "ReplenishDoneIn": {
        "properties": {
          "qty_moved": {
            "type": "integer",
            "title": "Qty Moved"
          }
        },
        "type": "object",
        "required": [
          "qty_moved"
        ],
        "title": "ReplenishDoneIn"
      },
      "ReplenishmentList": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/ReplenishmentTask"
            },
            "type": "array",
            "title": "Tasks"
          },
          "at_zero": {
            "type": "integer",
            "title": "At Zero",
            "description": "Pick faces with nothing left — these outrank the rest"
          }
        },
        "type": "object",
        "required": [
          "tasks",
          "at_zero"
        ],
        "title": "ReplenishmentList"
      },
...
      "RestockList": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/RestockRequest"
            },
            "type": "array",
            "title": "Requests"
          }
        },
        "type": "object",
        "required": [
          "requests"
        ],
        "title": "RestockList"
      },
...
      "ShortfallList": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ShortfallRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "by_person": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Person",
            "description": "Declarations per staffer — repeat offenders must be visible"
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "ShortfallList"
      },
...
      "Transfer": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "from_site_code": {
            "type": "string",
            "title": "From Site Code"
          },
          "to_site_code": {
            "type": "string",
            "title": "To Site Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dispatched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatched At"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "total_dispatched": {
            "type": "integer",
            "title": "Total Dispatched",
            "default": 0
          },
          "total_received": {
            "type": "integer",
            "title": "Total Received",
            "default": 0
          },
          "variance": {
            "type": "integer",
            "title": "Variance",
            "default": 0
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/TransferLine"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reference",
          "from_site_code",
          "to_site_code",
          "status"
        ],
        "title": "Transfer"
      },
      "TransferIn": {
        "properties": {
          "from_site_id": {
            "type": "integer",
            "title": "From Site Id"
          },
          "to_site_id": {
            "type": "integer",
            "title": "To Site Id"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/TransferLineIn"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "from_site_id",
          "to_site_id"
        ],
        "title": "TransferIn"
      },
...
      "TransferList": {
        "properties": {
          "transfers": {
            "items": {
              "$ref": "#/components/schemas/Transfer"
            },
            "type": "array",
            "title": "Transfers"
          }
        },
        "type": "object",
        "required": [
          "transfers"
        ],
        "title": "TransferList"
      },
...
    }
  }
}